Implemented peak merging
Implemented peak merging and writing results to file. Peak merging works by
checking whether or not two adjacent peaks are contained within a gene, or if
two adjacent peaks are within given threshold distance away from each other.
Results from the merging are written to a file named "merged.bed" The peaks
within a single chromosome are ordered (according to start position), but the
chromosomes themselves are not ordered.