Implemented geneID peak tracking
Peak scanning for within-gene peaks now associates the within-gene peaks with
one or more geneID's. This was done because one problem we ran into in the
previous implementation was that peak-gene membership would be ambiguous - one
peak would belong to overlapping genes. Also, the previous implementation was
linking within-gene peaks regardless of whether or not adjacent peaks were
actually within the same gene. Doing this kind of geneID tracking should fix
this problem.

Note: the .bed files in 6GapResults in this commit are associated with the
previous program version.