Implemented the arrowhead transformation
Currently, the transformation utilizes numpy ndarrays for all operations and
computations. The transformation involves computing the horizontal distance of
the current index to the diagonal, and then using that distance to access values
that are that distance away from the diagonal. Any distance for that goes out
of the matrix bounds is ignorred and the value in the transformed matrix is 0.