Completed implementation of DPRCS
Completed implementing the DP algorithm for summing contiguous elements in a
given array, for all "arrays" (rows/columns) of a given matrix. The result of
the pre-computation is a 3D array that stores the sum values, which can be
queried in constant time. The next step of the implementation will be to
introduce functions for the computation such that the same DP recurrence can be
used to compute different statistics for the given matrix (eg. S_sum, S_var,
and S_sign).