Variance Scoring Implementation
Implemented the triangle variance computation. In order to get the triangle
variances, create two UVTriangles (or LVTriangles) using the row/column sums -
one for the regular matrix, and one for the squared matrix. Call
regular.square(), and then do squared.subtract(regular). This will result in
squared having the results of the variance scores.