Fixed bug in Ssign Computation
The bug was in the fact that the base case was always being set to be the value
in the original grid, instead of that value's sign. This threw the rest of the
calculation off.

Another possible bug is that I currently consider 0's to be positive, which may
not be the right way to do things. Maybe this fix will come in a later commit.