Added a copy constructor to LTriangle
LTriangle needed a copy constructor because the original LTriangle matrix needs
to be used for the filtering operation. Previous code computed the score
matrices in the existing LTriangle object, so the original results were lost
in the computation of the score matrices. Output score matrices have been
confirmed to be unchanged from previous code.