RCS reimplementation
The previous ThreeD_Array reimplementation to fix the index too large issue
didn't work. It fixed the seg faults, but the run time and the space complexity
was way too much to be feasible (it consumed ~50GB memory before I manually
killed it).

This current implementation foregoes the precomputation in favor of computation
of the sums/signs on querying. I verified that the output was identical to the
previous implementation. However, there was a bug in the way I was computing
scores. Previously, I was taking the signs of both parameters, so the sign score
would never exceed [-2, 2]. I also did not account for 0's (I counted 0's as
positive). Now, I account for 0's and I only take the sign of the second input
parameter so the Ssign matrix should be accurate now.