Implemented Variable filtering parameters
Changed the filtering parameters from being fixed constants to variables that
can be varied by the filtering function caller (score.cpp). This was done to
allow us to create multiple filterings for the score matrix and to allow us to
select the best filtering. Currently, the thresholds passed to the filtering
function are identical to the constants previously, and the outputs from this
implementation are identical to the previous implementation.