Begun matlab domain calling
After filtering the score matrix, I use matlab to determine connected components
in the matrix. In this commit, I wrote some code (untested) to find all the
connected components, and to find the local maxima within each connected
component. The indices of the local maxima will be the domain boundaries.