Modified region chunking to account for reversed endpoints
Previously, the chunking process relied on the fact that the "start" of the
domain was numerically less than the "end" of the domain. However, if a domain's
endpoints were reversed (eg. end < start), the script would not chunk the
domain properly. This modified version handles backwards domains by determining
which endpoint is numerically less than the other one, and labeling that end
to be the start position.