Initial condense implementation
condense.py is a script that will take in a cumulative coverage file that
describes the coverage of several different markers divided into several
"merged" matrices and condense it into one matrix. The main goal of such a
script is to visualize the relationship of the various functional regions of
domains.